Dto refactor, small change in Profile
