--- src/main/java/com/myfinancemap/app/dto/UpdateProfileDto.java
@@ -1,12 +1,0 @@
-package com.myfinancemap.app.dto;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-public class UpdateProfileDto {
-    private String firstName;
-    private String lastName;
-    private LocalDateTime birthDate;
-}
