+++ src/main/java/com/myfinancemap/app/dto/profile/ProfileDto.java
@@ -1,14 +1,15 @@
-package com.myfinancemap.app.dto;
+package com.myfinancemap.app.dto.profile;
 
+import com.myfinancemap.app.dto.address.AddressDto;
 import lombok.Data;
 
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 
 @Data
 public class ProfileDto {
     private Long profileId;
     private String firstName;
     private String lastName;
-    private LocalDateTime birthDate;
+    private LocalDate birthDate;
     private AddressDto homeAddress;
 }
