+++ src/main/java/com/myfinancemap/app/dto/user/MinimalUserDto.java
@@ -1,4 +1,4 @@
-package com.myfinancemap.app.dto;
+package com.myfinancemap.app.dto.user;
 
 import lombok.Data;
