+++ src/main/java/com/myfinancemap/app/dto/user/UpdateUserDto.java
@@ -1,0 +1,9 @@
+package com.myfinancemap.app.dto.user;
+
+import com.myfinancemap.app.dto.profile.UpdateProfileDto;
+import lombok.Data;
+
+@Data
+public class UpdateUserDto {
+    private UpdateProfileDto profile;
+}
