+++ src/main/java/com/myfinancemap/app/dto/user/UserDto.java
@@ -1,5 +1,6 @@
-package com.myfinancemap.app.dto;
+package com.myfinancemap.app.dto.user;
 
+import com.myfinancemap.app.dto.profile.ProfileDto;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
