+++ src/main/java/com/myfinancemap/app/persistence/domain/Profile.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
+import java.time.LocalDate;
 
 @Entity
 @Table(name = "profile")
@@ -16,7 +16,7 @@ public class Profile {
     private Long profileId;
     private String firstName;
     private String lastName;
-    private LocalDateTime birthDate;
+    private LocalDate birthDate;
     @OneToOne
     @JoinColumn(name = "addressId", referencedColumnName = "addressId")
     private Address homeAddress;
